+++ libs/exception-handling/src/index.ts
@@ -1,4 +1,7 @@
 // Export all public APIs here
+export * from "./decorators/error-template.decorator"
 export * from "./exception-handler.module"
 export * from "./filters/exception.filter"
+export * from "./interceptors/error-template.interceptor"
 export * from "./interfaces/neoma-exception.interface"
+export * from "./pipes/validation.factory"